+++ src/components/auto-mode-button/DialogAutoMode.vue
